split example into two
